# Fix handling of summations with an empty range