--- a/src/macros.jl
+++ b/src/macros.jl
@@ -389,7 +389,7 @@ end
 function parse_one_operator_constraint(_error::Function, vectorized::Bool, sense::Val, lhs, rhs)
     # Simple comparison - move everything to the LHS.
     #
-    # Note: We add the +0 to this term to account for the diabolical case that
+    # Note: We add the +0 to this term to account for the pathological case that
     # the `lhs` is a `VariableRef` and the `rhs` is a summation with no terms.
     # Without the `+0` term, `aff` would evaluate to a `VariableRef` when we
     # really want it to be a `GenericAffExpr`.