Okay, so I thought of a few fixes for this, and I went for the easiest (if the most hacky).

This code should only ever get called for constraints with an (in)equality sense. So `a <= b`, `a >= b`, or `a == b`. (`a in b` goes through a different code path.) Given we have a comparison, then we always want to obtain a `GenericAffExpr`. The easiest way to do this is to just add a dummy `+ 0` term to the expression. 

I haven't benchmarked this, but it should be negligible.

The alternative is to only include this in the `parse_generator_sum` method, and maybe only if the generator has zero terms.

The other argument for doing what I've done here is that there might be other edge-cases (other than `sum` where it might occur. This approach is a panacea.

Closes #1954