@blegat 2019-05-07
If we always want it to be an expression and not a VariableRef, we can do
```julia
functionize(f::SingleVariable) = convert(AffExpr, f)
functionize(f::Any) = f
...
aff = :(functionize($lhs - $rhs))
```