@ccoffrin 2019-05-07
I would also recommend adding this test to avoid any regressions on the NL side,
```
    @testset "Empty summation in @NLconstraints" begin
        model = Model()
        @variable(model, x)
        c = @NLconstraint(model, x == sum(1.0 for i in 1:0))
    end
```