@odow 2019-05-07
So I've added the NL test. But going down @blegat's route got complicated since it makes the constraints parse differently.