@odow 2019-05-08
There is no change. 

Current master
```julia
Oscars-MBP:JuMP oscar$ git checkout master
Switched to branch 'master'
Your branch is up to date with 'origin/master'.
Oscars-MBP:JuMP oscar$ julia test/perf/speed.jl
P-Median(100 facilities, 100 customers, 5000 locations) benchmark:
BenchmarkTools.Trial: 
  memory estimate:  714.13 MiB
  allocs estimate:  12025585
  --------------
  minimum time:     1.491 s (51.07% GC)
  median time:      1.679 s (54.64% GC)
  mean time:        1.664 s (54.87% GC)
  maximum time:     1.808 s (58.46% GC)
  --------------
  samples:          4
  evals/sample:     1
Cont5(n=500) benchmark:
BenchmarkTools.Trial: 
  memory estimate:  402.78 MiB
  allocs estimate:  6785862
  --------------
  minimum time:     682.780 ms (37.28% GC)
  median time:      834.134 ms (49.49% GC)
  mean time:        797.730 ms (45.13% GC)
  maximum time:     900.495 ms (49.59% GC)
  --------------
  samples:          7
  evals/sample:     1
```

This PR
```Julia
Oscars-MBP:JuMP oscar$ git checkout od/fix_summation
Switched to branch 'od/fix_summation'
Your branch is up to date with 'origin/od/fix_summation'.
Oscars-MBP:JuMP oscar$ julia test/perf/speed.jl
P-Median(100 facilities, 100 customers, 5000 locations) benchmark:
BenchmarkTools.Trial: 
  memory estimate:  714.13 MiB
  allocs estimate:  12025585
  --------------
  minimum time:     1.484 s (51.12% GC)
  median time:      1.699 s (54.34% GC)
  mean time:        1.680 s (54.58% GC)
  maximum time:     1.838 s (57.81% GC)
  --------------
  samples:          4
  evals/sample:     1
Cont5(n=500) benchmark:
BenchmarkTools.Trial: 
  memory estimate:  402.78 MiB
  allocs estimate:  6785862
  --------------
  minimum time:     710.835 ms (37.76% GC)
  median time:      848.776 ms (49.96% GC)
  mean time:        809.624 ms (45.11% GC)
  maximum time:     894.192 ms (49.03% GC)
  --------------
  samples:          7
  evals/sample:     1
```