@blegat 2019-05-08
> There is no change.

Good! That means it is modifying the expression in-place by adding `0.0` to the constant which has negligible effect on performance.