@blegat 2019-05-08
Could you give the effect of this PR to the `speed.jl` benchmark ?